// src/game/models/GamePhase.ts
export enum GamePhase {
  SETUP = 'SETUP',
  TURN_START = 'TURN_START',
  ACTION_DECLARED = 'ACTION_DECLARED',
  CHALLENGE_WINDOW = 'CHALLENGE_WINDOW',
  BLOCK_WINDOW = 'BLOCK_WINDOW',
  BLOCK_CHALLENGE_WINDOW = 'BLOCK_CHALLENGE_WINDOW',
  ACTION_RESOLUTION = 'ACTION_RESOLUTION',
  SELECTING_CARD_TO_REVEAL = 'SELECTING_CARD_TO_REVEAL',
  SELECTING_CARDS_TO_EXCHANGE = 'SELECTING_CARDS_TO_EXCHANGE',
  EXCHANGING_CARDS = 'EXCHANGING_CARDS',
  TURN_END = 'TURN_END',
  GAME_OVER = 'GAME_OVER',
}